@@ -18,18 +18,17 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  modal_progress_hud: ^0.1.3
-  fluttertoast: ^4.0.1
-  webview_flutter: ^0.3.22+1
+  fluttertoast: ^8.0.8
+  webview_flutter: ^2.0.13
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3 
 
 dev_dependencies:
   flutter_test:
